+++ scm-ui/ui-components/src/CardColumn.tsx
@@ -27,7 +27,7 @@ import styled from "styled-components";
 import { Link } from "react-router-dom";
 
 type Props = {
-  title: string;
+  title: ReactNode;
   description?: string;
   avatar: ReactNode;
   contentRight?: ReactNode;
@@ -99,9 +99,7 @@ export default class CardColumn extends React.Component<Props> {
           <FlexFullHeight className={classNames("media-content", "text-box", "is-flex")}>
             <div className="is-flex">
               <ContentLeft className="content">
-                <p className="shorten-text is-marginless">
-                  <strong>{title}</strong>
-                </p>
+                <p className="shorten-text is-marginless">{title}</p>
                 <p className="shorten-text">{description}</p>
               </ContentLeft>
               <ContentRight>{contentRight}</ContentRight>