+++ scm-ui/ui-components/src/forms/Checkbox.stories.tsx
@@ -42,4 +42,3 @@ storiesOf("Forms|Checkbox", module)
       <Checkbox label="Checked but disabled" checked={true} disabled={true} />
     </Spacing>
   ));
-